+++ src/main/java/pl/cyfronet/fid/cmdbproxy/web/filter/CmdbCrudCredentialsFilter.java
@@ -19,7 +19,7 @@ import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 @Component
-@Order(5)
+@Order(6)
 public class CmdbCrudCredentialsFilter extends CmdbCrudAwareFilter {
 
     @Value("${proxy.cmdb-crud.username}")
@@ -80,7 +80,7 @@ public class CmdbCrudCredentialsFilter extends CmdbCrudAwareFilter {
                 }
             }
 
-            if (cmdbCrudRequest((HttpServletRequest)getRequest())) {
+            if (isCmdbCrudRequest((HttpServletRequest)getRequest())) {
                 headerNames.add(AUTHORIZATION_HEADER_NAME);
             }